This is the first commit
Here we have the changes made in angular.json and package.json
that are necessary for building the project.
Since it is the first commit, it contains all the changes made in these
files from the begin.
So far in this project we have setup the topbar,sidemenu and the
content page but without styles fully applied