This is the second commit for the dashbord2
Here we have the changes made in .gitignore and polyfills
but actually it those changes have no real impact on the project's
state.
The current commit is done just by tracking purpose.
So far in this project we have setup the topbar,sidemenu and the content
page but without styles fully applied.